rename generic 'bearer strategy' into specific 'azure ad strategy'
